% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_marked_value.R
\name{calculate_marked_value}
\alias{calculate_marked_value}
\title{Calculate marked value}
\usage{
calculate_marked_value(
  df_portfolio,
  etf_info,
  df_closing_price,
  benchmark_name = "iShares MSCI ACWI UCITS ETF",
  benchmark_ticker = "IUSQ.DE",
  performance_info
)
}
\arguments{
\item{df_portfolio}{data.frame with stocks}

\item{etf_info}{list with ETF info}

\item{df_closing_price}{data.frame with Date and and a column for each stock 
with closing price in DKK}

\item{benchmark_name}{character string with benchmark name}

\item{benchmark_ticker}{character string with benchmark ticker}

\item{performance_info}{list with performance info (sector and geography)}
}
\description{
Calculate marked value
}
