% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_error.R
\name{catch_error}
\alias{catch_error}
\title{An extension of tryCatch}
\usage{
catch_error(expr)
}
\arguments{
\item{expr}{expression to evalute}
}
\value{
A list with value, warning and error message
}
\description{
An extension of tryCatch
}
