% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_in_db.R
\name{check_file_in_db}
\alias{check_file_in_db}
\title{Check file in database}
\usage{
check_file_in_db(file, max_days)
}
\arguments{
\item{file}{path to file}

\item{max_days}{integer indicating maximum number days allowed before new data load is needed}
}
\description{
Check file in database
}
