% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_with_benchmark_portfolio.R
\name{compare_with_benchmark_portfolio}
\alias{compare_with_benchmark_portfolio}
\title{Compare with benchmark portfolio}
\usage{
compare_with_benchmark_portfolio(
  df_holdings,
  benchmark_info,
  var,
  df_performance
)
}
\arguments{
\item{df_holdings}{data.frame with holdings for your portfolio}

\item{benchmark_info}{benchmark information obtained from load_benchmark_info}

\item{var}{variable name (string) used to compare your portfolio with benchmark}

\item{df_performance}{data.frame with performance statistics}
}
\value{
Summarised data.frame with marked value proportion grouped by `var`
}
\description{
Compare with benchmark portfolio
}
