% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_scandinavian_letters.R
\name{contains_scandinavian_letters}
\alias{contains_scandinavian_letters}
\title{Check if vector contains scandinavian letters}
\usage{
contains_scandinavian_letters(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Check if vector contains scandinavian letters
}
