% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data_frame_to_vector.R
\name{convert_data_frame_to_vector}
\alias{convert_data_frame_to_vector}
\title{Convert data frame to vector}
\usage{
convert_data_frame_to_vector(data)
}
\arguments{
\item{data}{data.frame}
}
\description{
Convert data frame to vector
}
