% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_dkk_with_constant.R
\name{convert_to_dkk_with_constant}
\alias{convert_to_dkk_with_constant}
\title{Convert to dkk with constant}
\usage{
convert_to_dkk_with_constant(x, currency)
}
\arguments{
\item{x}{numeric vector}

\item{currency}{currency to convert from ("USD", "EUR" or "SEK")}
}
\description{
Convert to dkk with constant
}
