% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_load_data.R
\name{db_load_data}
\alias{db_load_data}
\title{Database load data}
\usage{
db_load_data(ticker, from, folder_name = "stocks", max_days = 3)
}
\arguments{
\item{ticker}{ticker code}

\item{from}{from date}

\item{folder_name}{name of folder used as database}

\item{max_days}{max days since last load (default: 3)}
}
\description{
Database load data
}
