% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_load_stock_price.R
\name{db_load_stock_price}
\alias{db_load_stock_price}
\title{Database load stock price}
\usage{
db_load_stock_price(ticker, folder_name = "stocks")
}
\arguments{
\item{ticker}{ticker code}

\item{folder_name}{name of folder used as database}
}
\description{
Database load stock price
}
