% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_sector_performance.R
\name{db_sector_performance}
\alias{db_sector_performance}
\title{Database sector performance}
\usage{
db_sector_performance(
  updateProgress = NULL,
  folder_name = "performance",
  max_days = 3
)
}
\arguments{
\item{updateProgress}{updateProgress function}

\item{folder_name}{name of folder (default is performance)}

\item{max_days}{integer specifying maximum number of days before data is considered too old}
}
\description{
Database sector performance
}
