% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_portfolio.R
\name{delete_it_all}
\alias{delete_it_all}
\title{Delete it all}
\usage{
delete_it_all(df_portfolio, rows_selected)
}
\arguments{
\item{df_portfolio}{data.frame with stocks}

\item{rows_selected}{integer vector with rows to delete}
}
\value{
logical, TRUE if entire portfolio should be deleted
}
\description{
Delete it all
}
