% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_portfolio.R
\name{delete_rows_from_portfolio}
\alias{delete_rows_from_portfolio}
\title{Delete rows from portfolio}
\usage{
delete_rows_from_portfolio(
  df_portfolio,
  df_closing_price,
  etf_info,
  rows_selected
)
}
\arguments{
\item{df_portfolio}{data.frame with stocks}

\item{df_closing_price}{data.frame with Date and closing price in DKK for each stock}

\item{etf_info}{list with ETF info (NULL for stocks)}

\item{rows_selected}{index vector corresponding to rows to delete}
}
\description{
Delete rows from portfolio
}
