% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_source.R
\name{get_danske_source}
\alias{get_danske_source}
\title{Get Danske Invest source}
\usage{
get_danske_source()
}
\description{
Get Danske Invest source
}
\details{
Ticker is used to get marked value today
Ticker_approx is not used in the code. It is only used to keep track of which ETF is used as approximation when loading holding analytics from iShares
iShares_url is used to get holdings and sector information (url is found under Holdings table: Detailed Holdings and Analytics)
Skip is the parameter used in read.csv for the specific url (Skipping rows before header appear)
}
