% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_etf_info.R
\name{get_etf_info}
\alias{get_etf_info}
\title{Get ETF info}
\usage{
get_etf_info(stock_name, ticker, max_days = 3)
}
\arguments{
\item{stock_name}{stock name}

\item{ticker}{ticker code}
}
\description{
Get ETF info
}
