% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_and_stock_type_performance.R
\name{get_geo_and_stock_type_performance}
\alias{get_geo_and_stock_type_performance}
\title{Get performance}
\usage{
get_geo_and_stock_type_performance()
}
\description{
Get performance
}
