% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_holdings.R
\name{get_holdings}
\alias{get_holdings}
\title{Get holdings}
\usage{
get_holdings(df_portfolio, etf_info)
}
\arguments{
\item{df_portfolio}{df_portfolio}

\item{eft_info}{etf_info}
}
\description{
Get holdings
}
