% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_etf_info.R
\name{get_holdings_info}
\alias{get_holdings_info}
\title{Get holdings info}
\usage{
get_holdings_info(data, ticker, top_n = 100)
}
\arguments{
\item{data}{data.frame from load_etf_info}

\item{top_n}{number of rows to return. Default is top 100}
}
\description{
Get holdings info
}
