% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marked_value_today.R
\name{get_market_value_today}
\alias{get_market_value_today}
\title{Get marked value today}
\usage{
get_market_value_today(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Get marked value today
}
