% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_change.R
\name{get_price_change.data.frame}
\alias{get_price_change.data.frame}
\title{Method for getting price change for data.frame}
\usage{
\method{get_price_change}{data.frame}(obj, var_name)
}
\arguments{
\item{obj}{object}
}
\description{
Method for getting price change for data.frame
}
