% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_change.R
\name{get_price_change.integer}
\alias{get_price_change.integer}
\title{Method for getting price change for integer vector}
\usage{
\method{get_price_change}{integer}(obj)
}
\arguments{
\item{obj}{object}
}
\description{
Method for getting price change for integer vector
}
