% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_change.R
\name{get_price_change.numeric}
\alias{get_price_change.numeric}
\title{Method for getting price change for numeric vector}
\usage{
\method{get_price_change}{numeric}(obj)
}
\arguments{
\item{obj}{object}
}
\description{
Method for getting price change for numeric vector
}
