% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_change.R
\name{get_price_change}
\alias{get_price_change}
\title{Get price change}
\usage{
get_price_change(obj, ...)
}
\arguments{
\item{obj}{object to calculate price change on}
}
\description{
Get price change
}
