% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sector_performance.R
\name{get_sector_performance}
\alias{get_sector_performance}
\title{Get sector performance}
\usage{
get_sector_performance(updateProgress = NULL)
}
\arguments{
\item{updateProgress}{updateProgress function}
}
\description{
Get sector performance
}
