% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_source.R
\name{get_single_stock_source}
\alias{get_single_stock_source}
\title{Get single stock source}
\usage{
get_single_stock_source()
}
\description{
Single stock information regarding stock name and ticker code
}
\details{
Add stocks by searching for the name on iShares and find the ticker code
If a new source category should be added, remember to include the new function to get_stock_name_and_ticker
on yahoo
}
