% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_source.R
\name{get_sparindex_source}
\alias{get_sparindex_source}
\title{Get sparindex source}
\usage{
get_sparindex_source()
}
\description{
Get sparindex source
}
\details{
Ticker is used to get marked value today
Ticker_approx is not used in the code. It is only used to keep track of which ETF is used as approximation when loading holding analytics from iShares
iShares_url is used to get holdings and sector information
Skip is the parameter used in read.csv for the specific url
}
