% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stock_name.R
\name{get_stock_name}
\alias{get_stock_name}
\title{Get stock name}
\usage{
get_stock_name(ticker, df, shiny = FALSE)
}
\arguments{
\item{ticker}{ticker code}

\item{df}{data.frame stored using readRDS(file = "./data/df_stock_name_and_ticker.RDS")}

\item{shiny}{logical indicating if the first element should be blank}
}
\description{
Get stock name
}
