% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stock_name_and_ticker.R
\name{get_stock_name_and_ticker_danske}
\alias{get_stock_name_and_ticker_danske}
\title{Get stock name and ticker for danske invest}
\usage{
get_stock_name_and_ticker_danske()
}
\description{
Get stock name and ticker for danske invest
}
