% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stock_name_and_ticker.R
\name{get_stock_name_and_ticker_ishares}
\alias{get_stock_name_and_ticker_ishares}
\title{Get stock name and ticker for iShares products}
\usage{
get_stock_name_and_ticker_ishares()
}
\description{
Get stock name and ticker for iShares products
}
