% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stock_name_and_ticker.R
\name{get_stock_name_and_ticker_single_stocks}
\alias{get_stock_name_and_ticker_single_stocks}
\title{Get stock name and ticker for single stocks}
\usage{
get_stock_name_and_ticker_single_stocks()
}
\description{
Get stock name and ticker for single stocks
}
