% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stock_name_and_ticker.R
\name{get_stock_name_and_ticker_sparindex}
\alias{get_stock_name_and_ticker_sparindex}
\title{Get stock name and ticker for Sparindex products}
\usage{
get_stock_name_and_ticker_sparindex()
}
\description{
Get stock name and ticker for Sparindex products
}
