% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stock_name_and_ticker.R
\name{get_stock_name_and_ticker_usa}
\alias{get_stock_name_and_ticker_usa}
\title{Get stock name and ticker for American stocks}
\usage{
get_stock_name_and_ticker_usa()
}
\description{
Get stock name and ticker for American stocks
}
