% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ticker.R
\name{get_ticker}
\alias{get_ticker}
\title{Get ticker}
\usage{
get_ticker(stock_name, df, shiny = FALSE)
}
\arguments{
\item{stock_name}{Stock name}

\item{df}{data.frame stored using readRDS(file = "./data/df_stock_name_and_ticker.RDS")}

\item{shiny}{logical indicating if the first element should be blank}
}
\description{
Get ticker
}
