% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_benchmark_info.R
\name{load_benchmark_info}
\alias{load_benchmark_info}
\title{Load benchmark info}
\usage{
load_benchmark_info(df_portfolio, etf_info)
}
\arguments{
\item{df_portfolio}{df_portfolio}

\item{etf_info}{etf_info obtained from get_etf_info}
}
\description{
Load benchmark info
}
