% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_portfolio.R
\name{load_portfolio}
\alias{load_portfolio}
\title{Load portfolio}
\usage{
load_portfolio(file_name, overwrite_portfolio, df)
}
\arguments{
\item{file_name}{name of file placed in data/portfolios folder}

\item{overwrite_portfolio}{logical (TRUE/FALSE)}

\item{df}{data.frame with existing portfolio (only relevant if overwrite_portfolio = FALSE)}
}
\description{
Load portfolio
}
