% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_country_to_currency.R
\name{map_country_to_currency}
\alias{map_country_to_currency}
\title{Map country to currency}
\usage{
map_country_to_currency(x, currency = rep(NA, length(x)))
}
\arguments{
\item{x}{character vector with countries}

\item{currency}{optional, if currency is supplied then the currency is simply returned}
}
\description{
Map country to currency
}
