% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_country_to_region.R
\name{map_country_to_region}
\alias{map_country_to_region}
\title{Map country to region}
\usage{
map_country_to_region(x)
}
\arguments{
\item{country}{character vector with countries}
}
\description{
Map country to region
}
