% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_and_stock_type_performance.R
\name{map_stock_name_to_region}
\alias{map_stock_name_to_region}
\title{Map stock name to region}
\usage{
map_stock_name_to_region(x)
}
\arguments{
\item{x}{character string with sparindex stock}
}
\description{
Map stock name to region
}
