% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_and_stock_type_performance.R
\name{map_stock_name_to_stock_type}
\alias{map_stock_name_to_stock_type}
\title{Map stock name to stock type}
\usage{
map_stock_name_to_stock_type(x)
}
\arguments{
\item{x}{character string with sparindex stock}
}
\description{
Map stock name to stock type
}
