% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pie_chart.R
\name{plot_pie_chart}
\alias{plot_pie_chart}
\title{Plot pie chart}
\usage{
plot_pie_chart(data, labels)
}
\arguments{
\item{data}{data.frame}

\item{labels}{labels as string}
}
\description{
Plot pie chart
}
