% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_on_stock_input.R
\name{react_on_ticker_input}
\alias{react_on_ticker_input}
\title{React on ticker input}
\usage{
react_on_ticker_input(df_stocks, stock_name, ticker)
}
\arguments{
\item{df_stocks}{data.frame with auto filled stock name and ticker codes}

\item{stock_name}{name of stock}

\item{ticker}{ticker code}
}
\value{
character string with stock name based on ticker input
}
\description{
React on ticker input
}
