% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_panel_analytics.R
\name{render_panel_analytics}
\alias{render_panel_analytics}
\title{Render panel analytics}
\usage{
render_panel_analytics(ns)
}
\arguments{
\item{ns}{namespace}
}
\description{
Render panel analytics
}
