% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_scandinavian_letters.R
\name{replace_scandinavian_letters}
\alias{replace_scandinavian_letters}
\title{Replace scandinavian letters}
\usage{
replace_scandinavian_letters(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Replace scandinavian letters
}
