% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_portfolio.R
\name{save_portfolio}
\alias{save_portfolio}
\title{Save portfolio}
\usage{
save_portfolio(df_portfolio, id)
}
\arguments{
\item{df_portfolio}{data.frame to save}

\item{id}{id used to uniquely identify row in portfolio log}
}
\description{
Save portfolio
}
