% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_panel_analytics.R
\name{show_panel_analytics}
\alias{show_panel_analytics}
\title{Show panel analytics}
\usage{
show_panel_analytics(title, outputId)
}
\arguments{
\item{title}{title of panel}

\item{outputId}{outputId of plotly object}
}
\description{
Show panel analytics
}
