% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_version_number.R
\name{show_version_number}
\alias{show_version_number}
\title{Show version number}
\usage{
show_version_number(package = "shinyStocks")
}
\arguments{
\item{package}{package name}
}
\description{
Show version number
}
