% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringify.R
\name{stringify}
\alias{stringify}
\title{Stringify}
\usage{
stringify(x)
}
\arguments{
\item{x}{character vector with stock name or ticker that should be modified into a string not having any dots, spaces or dashes}
}
\description{
Stringify
}
