% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_log.R
\name{update_log}
\alias{update_log}
\title{Update log}
\usage{
update_log(df_log, id, new_date)
}
\arguments{
\item{df_log}{data.frame with log table}

\item{id}{row that should be updated identified using id}

\item{new_date}{the new date used when updating the date column}
}
\description{
Update log
}
