% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorized_convert_to_dkk.R
\name{vectorized_convert_to_dkk}
\alias{vectorized_convert_to_dkk}
\title{Vectorized convert to dkk}
\usage{
vectorized_convert_to_dkk(from)
}
\arguments{
\item{from}{character vector with currencies}
}
\description{
Vectorized convert to dkk
}
