% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorized_load_data_and_get_etf_info.R
\name{vectorized_load_data_and_get_etf_info}
\alias{vectorized_load_data_and_get_etf_info}
\title{Vectorized load data and get etf info}
\usage{
vectorized_load_data_and_get_etf_info(
  ticker,
  stock_name = rep(NA, length(ticker)),
  from,
  rate = rep(1, length(ticker)),
  updateProgress = NULL
)
}
\arguments{
\item{ticker}{ticker code}

\item{stock_name}{optional. Used to asses if load function for ETFs should be called.}

\item{from}{from date}

\item{rate}{rate which is used to calculate closing price in DKK}

\item{updateProgress}{progress function}
}
\description{
Vectorized load data and get etf info
}
