% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_log.R
\name{write_to_log}
\alias{write_to_log}
\title{Write to log}
\usage{
write_to_log(user_name, portfolio_name, id)
}
\arguments{
\item{user_name}{character string with user name}

\item{id}{id used to uniquely identify row in portfolio log}

\item{character}{sring with portfolio name}
}
\description{
Write to log
}
