% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_marked_value.R
\name{calculate_marked_value}
\alias{calculate_marked_value}
\title{Calculate marked value}
\usage{
calculate_marked_value(
  df_portfolio,
  etf_info,
  df_marked_value,
  benchmark_name = "iShares MSCI ACWI UCITS ETF",
  benchmark_ticker = "IUSQ.DE"
)
}
\arguments{
\item{df_portfolio}{data.frame with stocks}

\item{etf_info}{list with ETF info}

\item{df_marked_value}{data.frame with marked value column for each stock}

\item{benchmark_name}{character string with benchmark name}

\item{benchmark_ticker}{character string with benchmark ticker}
}
\description{
Calculate marked value
}
