% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_load_data.R
\name{db_load_data}
\alias{db_load_data}
\title{Database load data}
\usage{
db_load_data(ticker, from, folder_name = "stocks")
}
\arguments{
\item{ticker}{ticker code}

\item{from}{from date}

\item{folder_name}{name of folder used as database}
}
\description{
Database load data
}
