% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_country_to_currency.R
\name{map_country_to_currency}
\alias{map_country_to_currency}
\title{Map country to currency}
\usage{
map_country_to_currency(x)
}
\arguments{
\item{x}{character vector with countries}
}
\description{
Map country to currency
}
