% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_etf_info.R
\name{save_etf_info}
\alias{save_etf_info}
\title{Save ETF info}
\usage{
save_etf_info(obj, stock_name)
}
\arguments{
\item{obj}{object to save}

\item{stock_name}{stock name}
}
\description{
Save ETF info
}
