% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorized_load_data.R
\name{vectorized_load_data}
\alias{vectorized_load_data}
\title{Vectorized load data}
\usage{
vectorized_load_data(
  ticker,
  from,
  cbind = TRUE,
  rate = rep(1, length(ticker)),
  updateProgress = NULL
)
}
\arguments{
\item{ticker}{ticker code}

\item{from}{from date}

\item{cbind}{logical indicating if output should column binded}

\item{rate}{currency conversion rate to DKK. Use rate = 1 if no conversion should be applied}

\item{updateProgress}{progress function}
}
\description{
Vectorized load data
}
